[MINOR] refactor: Rename MultiStorage to HybridStorage
