+++ server/src/test/java/org/apache/uniffle/server/ShuffleFlushManagerTest.java
@@ -58,7 +58,7 @@ import org.apache.uniffle.server.event.AppPurgeEvent;
 import org.apache.uniffle.server.storage.HadoopStorageManager;
 import org.apache.uniffle.server.storage.LocalStorageManager;
 import org.apache.uniffle.server.storage.LocalStorageManagerFallbackStrategy;
-import org.apache.uniffle.server.storage.MultiStorageManager;
+import org.apache.uniffle.server.storage.HybridStorageManager;
 import org.apache.uniffle.server.storage.StorageManager;
 import org.apache.uniffle.server.storage.StorageManagerFactory;
 import org.apache.uniffle.storage.HadoopTestBase;
@@ -646,7 +646,7 @@ public class ShuffleFlushManagerTest extends HadoopTestBase {
   }
 
   @Test
-  public void fallbackWrittenWhenMultiStorageManagerEnableTest(@TempDir File tempDir)
+  public void fallbackWrittenWhenHybridStorageManagerEnableTest(@TempDir File tempDir)
       throws InterruptedException {
     shuffleServerConf.setLong(ShuffleServerConf.FLUSH_COLD_STORAGE_THRESHOLD_SIZE, 10000L);
     shuffleServerConf.setString(
@@ -655,7 +655,7 @@ public class ShuffleFlushManagerTest extends HadoopTestBase {
         RssBaseConf.RSS_STORAGE_BASE_PATH, Arrays.asList(tempDir.getAbsolutePath()));
     shuffleServerConf.set(ShuffleServerConf.DISK_CAPACITY, 100L);
     shuffleServerConf.setString(
-        ShuffleServerConf.MULTISTORAGE_FALLBACK_STRATEGY_CLASS,
+        ShuffleServerConf.HYBRID_STORAGE_FALLBACK_STRATEGY_CLASS,
         LocalStorageManagerFallbackStrategy.class.getCanonicalName());
 
     StorageManager storageManager =
@@ -685,8 +685,8 @@ public class ShuffleFlushManagerTest extends HadoopTestBase {
     ShuffleDataFlushEvent bigEvent =
         new ShuffleDataFlushEvent(1, "1", 1, 1, 1, 100, blocks, null, null);
     bigEvent.setUnderStorage(
-        ((MultiStorageManager) storageManager).getWarmStorageManager().selectStorage(event));
-    ((MultiStorageManager) storageManager).getWarmStorageManager().updateWriteMetrics(bigEvent, 0);
+        ((HybridStorageManager) storageManager).getWarmStorageManager().selectStorage(event));
+    ((HybridStorageManager) storageManager).getWarmStorageManager().updateWriteMetrics(bigEvent, 0);
 
     event = createShuffleDataFlushEvent(appId, 1, 1, 1, null, 100);
     flushManager.addToFlushQueue(event);
@@ -706,7 +706,7 @@ public class ShuffleFlushManagerTest extends HadoopTestBase {
     shuffleServerConf.set(ShuffleServerConf.SERVER_FLUSH_HADOOP_THREAD_POOL_SIZE, 1);
     shuffleServerConf.set(ShuffleServerConf.SERVER_FLUSH_LOCALFILE_THREAD_POOL_SIZE, 1);
     shuffleServerConf.setString(
-        ShuffleServerConf.MULTISTORAGE_FALLBACK_STRATEGY_CLASS,
+        ShuffleServerConf.HYBRID_STORAGE_FALLBACK_STRATEGY_CLASS,
         LocalStorageManagerFallbackStrategy.class.getCanonicalName());
 
     StorageManager storageManager =
@@ -739,8 +739,8 @@ public class ShuffleFlushManagerTest extends HadoopTestBase {
     ShuffleDataFlushEvent bigEvent =
         new ShuffleDataFlushEvent(1, "1", 1, 1, 1, 100, blocks, null, null);
     bigEvent.setUnderStorage(
-        ((MultiStorageManager) storageManager).getWarmStorageManager().selectStorage(event));
-    ((MultiStorageManager) storageManager).getWarmStorageManager().updateWriteMetrics(bigEvent, 0);
+        ((HybridStorageManager) storageManager).getWarmStorageManager().selectStorage(event));
+    ((HybridStorageManager) storageManager).getWarmStorageManager().updateWriteMetrics(bigEvent, 0);
 
     event = createShuffleDataFlushEvent(appId, 1, 1, 1, null, 100);
     flushManager.addToFlushQueue(event);
